PUBDEV-4826: deeplearning mojo.  Added runit test and enable R clients to call deeplearning mojo. PUBDEV-4826: Deeplearning mojo.  Fixed effect of distribution on regression output. PUBDEV-4826: Finished autoencoder mojo. PUBDEV-4826: added Runit test. PUBDEV-4826: Fixed python test harness. PUBDEV-4826L Fixed R test harness. PUBDEV-4628: Deeplearning mojo: Change neural network inputs to double, weights to float[] and keep bias as double[] just like the original deeplearning toolbox.  Accuracy is now achieved to be the same as deeplearning pojo.

+++ h2o-genmodel/src/main/java/hex/genmodel/algos/deeplearning/ActivationUtils.java
@@ -45,7 +45,7 @@ public class ActivationUtils {
   public static class ExpRectifierDropoutOut extends ExpRectifierOut {
     public double[] eval(double[] input, double drop_out_ratio, int maxOutk) {
       double[] output = super.eval(input, drop_out_ratio, maxOutk);
-      applyDropout(output, (float) drop_out_ratio, input.length);
+      applyDropout(output, drop_out_ratio, input.length);
       return output;
     }
   }
@@ -133,7 +133,7 @@ public class ActivationUtils {
       double[] output = new double[nodeSize];
 
       for (int index=0; index < nodeSize; index++)
-        output[index] = 1-2/(1+(float)Math.exp(2*input[index]));
+        output[index] = 1-2/(1+Math.exp(2*input[index]));
 
       return output;
     }
